mm: fix ia16 page-pool base truncation through uintptr_t
3693c0be moved the runtime page-pool base from a direct PAGE_POOL_BASE
return into a file-local static, then back through page_pool_base().
The static was typed uintptr_t — on every supported target except ia16
that is a 32-bit type and the change is invisible.

On ia16 small model uintptr_t is the 16-bit near-pointer type.  The
assignment

  static uintptr_t s_pool_base;
  s_pool_base = PAGE_POOL_BASE;     /* i16_page_pool_base, uint32_t */

silently truncates the upper 16 bits, so a pool base of 0x2a000 (set
correctly by stage2 → seg_init_modules) becomes 0x0a000 in s_pool_base.
mm_init then seeds the page allocator with addresses inside the VFS
code / data segments.  The first user-page write corrupts an
in-flight far-call frame, the CPU triple-faults, BIOS POST reboots,
and target_pcxt's breadcrumb check fires "warm reboot (kernel
re-entered without BIOS POST)" on the second boot.

Cure the truncation by typing the pool base and the mem_helper hook
that sets it as uint32_t.  Linear page-pool addresses can sit above
64 KB on ia16, so uintptr_t was always the wrong type here.

* s_pool_base: uintptr_t -> uint32_t
* mem_helper_init_pool(uint32_t *base_out)
* Xtensa override signature updated to match (no behaviour change,
  uintptr_t on xtensa is already 32-bit)

Verified with ./scripts/run.sh --test:
  qemu_arm    24/24 pass
  qemu_m68k   24/24 pass
  qemu_rv32   23/23 pass
  pcxt --hdd  18/18 pass

+++ src/arch/xtensa/kernel/core/mem_helper.c
@@ -278,7 +278,7 @@ int mem_helper_init_arenas(void) {
   return 0;
 }
 
-int mem_helper_init_pool(uintptr_t *base_out) {
+int mem_helper_init_pool(uint32_t *base_out) {
   const uint32_t caps = MALLOC_CAP_8BIT | MALLOC_CAP_INTERNAL;
   uint32_t free_before = (uint32_t)heap_caps_get_free_size(caps);
   uint32_t want = PAGE_COUNT_MAX;
